backend_neural/requirements.txt
flask
PyYaml
ftp_downloader
python-dotenv

tensorflow
numpy
matplotlib
opencv-python
IPython
scikit-image
ipykernel